fixed last bugs
