+++ hw4/moduleSort/qsortTests.c
@@ -1,11 +1,40 @@
 #include "qsortTests.h"
 #include <stdbool.h>
+#include <stdlib.h>
+#include "myQsort.h"
 
 #define ARRAY_SIZE 20
 
 int comparator(const int* first, const int* second)
 {
     return *first - *second;
+}
+
+void frequentElementCheck(int* inputArray, int arraySize, int* maximumCounter, int* frequentElement)
+{
+    int currentCounter = 1;
+    *maximumCounter = 0;
+    for (int i = 1; i < arraySize; ++i)
+    {
+        if (inputArray[i - 1] != inputArray[i])
+        {
+            if (currentCounter > *maximumCounter)
+            {
+                *maximumCounter = currentCounter;
+                currentCounter = 1;
+                *frequentElement = inputArray[i - 1];
+            }
+        }
+        else
+        {
+            ++currentCounter;
+        }
+    }
+    if (currentCounter > *maximumCounter)
+    {
+        *maximumCounter = currentCounter;
+        *frequentElement = inputArray[arraySize - 1];
+    }
 }
 
 bool qsortCheck(int* sortingArray, int* correctArray, int arraySize)
