+++ hw4/moduleSort/qsortTests.h
@@ -2,8 +2,8 @@
 
 #include <stdbool.h>
 
-//This function checks if qsort sorts the array correctly
+// This function checks if qsort sorts the array correctly
 bool qsortCheck(int* sortingArray, int* correctArray, int arraySize);
 
-//This function checks if test executed without errors
+// This function checks if test executed without errors
 bool testCorrectCase();
